video-translator/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
google-genai
elevenlabs
pysubs2
pydub
python-dotenv
